Migrated Simple consumer.
